Updated engine disposals throughout, hopefully decreasing database connections
