+++ Base/character_functions.py
