+++ utils/Char_Getter.py
@@ -6,7 +6,7 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import sessionmaker
 
 from EPF.EPF_Character import get_EPF_Character
-from Generic.Character import Character
+from Base.Character import Character
 from database_operations import USERNAME, PASSWORD, HOSTNAME, PORT, SERVER_DATA
 from database_models import get_tracker
 from database_operations import get_asyncio_db_engine
